Completed assignment

+++ Binomial.java
@@ -11,7 +11,7 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, basic version.
-	public static int binomial1(int n, int k) { 
+	public static long binomial1(int n, int k) { 
 		
      if (k > n) return 0;
      if (k == 0 || n == 0) return 1;
@@ -25,9 +25,9 @@ public class Binomial {
 		
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 		
-		int [][] nemo = new int [n + 1][k+1];
+		long [][] nemo = new long [n + 1][k+1];
 		for( int i = 0; i < nemo.length; i++) {
               for ( int j = 0; j < nemo[0].length; j++) {
